@@ -14,10 +14,10 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
-  provider: 3.2.0
+  provider: 4.3.3
   flutter:
     sdk: flutter
   intl: ^0.15.8
