@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
 
-class Produto with ChangeNotifier{
+class Product with ChangeNotifier {
   final String id;
   final String titulo;
   final String descricao;
@@ -8,14 +8,13 @@ class Produto with ChangeNotifier{
   final String urlDaImagem;
   bool ehFavorito;
 
-  Produto({
-    @required this.id,
-    @required this.titulo,
-    @required this.descricao,
-    @required this.preco,
-    @required this.urlDaImagem,
-    this.ehFavorito = false
-  });
+  Product(
+      {@required this.id,
+      @required this.titulo,
+      @required this.descricao,
+      @required this.preco,
+      @required this.urlDaImagem,
+      this.ehFavorito = false});
 
   void marcarComoFavorito() {
     ehFavorito = !ehFavorito;
